Fix segfaults during point matching.
These segfaults were all due to unexpected data -- e.g. images with no
keypoints or keypoints with fewer than two coarse correspondences in the
other image.

I also found a segfault during arrangement construction. This was due
not to a bug in the code, but to an empty image file, which I removed
(image 45).